deleted unused files, deleted idle status
