--- portal/middleware/django_online_status/.hgignore
@@ -1,21 +1,0 @@
-syntax: glob
-
-*.elc
-*.pyc
-*.db
-glob:django.wsgi
-
-syntax: regexp
-^manage\.py$
-^settings\.py$
-^syncdb\.py$
-^urls\.py$
-^templates$
-^.settings$
-^\.project$
-^\.pydevproject$
-
-syntax: regexp
-^django_online_status\.egg-info$
-^dist$
-^build$
