--- portal/middleware/django_online_status/AUTHORS
@@ -1,1 +1,0 @@
-Jakub Zalewski http://zalew.tumblr.com
