--- portal/middleware/django_online_status/django_online_status.egg-info/PKG-INFO
@@ -1,17 +1,0 @@
-Metadata-Version: 1.1
-Name: django-online-status
-Version: 0.1.0
-Summary: UNKNOWN
-Home-page: http://bitbucket.org/zalew/django-online-status
-Author: Jakub Zalewski
-Author-email: zalew7@gmail.com
-License: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
-Classifier: Development Status :: Beta
-Classifier: Environment :: Web Environment
-Classifier: Intended Audience :: Developers
-Classifier: License :: OSI Approved :: BSD License
-Classifier: Operating System :: OS Independent
-Classifier: Programming Language :: Python
-Classifier: Framework :: Django
