--- portal/middleware/django_online_status/django_online_status.egg-info/SOURCES.txt
@@ -1,17 +1,0 @@
-README
-setup.py
-django_online_status.egg-info/PKG-INFO
-django_online_status.egg-info/SOURCES.txt
-django_online_status.egg-info/dependency_links.txt
-django_online_status.egg-info/top_level.txt
-online_status/__init__.py
-online_status/middleware.py
-online_status/models.py
-online_status/status.py
-online_status/tests.py
-online_status/urls.py
-online_status/utils.py
-online_status/views.py
-online_status/templatetags/__init__.py
-online_status/templatetags/online_status_tags.py
-online_status/templatetags/status_format.py
