--- portal/middleware/django_online_status/django_online_status.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-online_status
