--- portal/middleware/django_online_status/locale/zh_TW/LC_MESSAGES/django.po
@@ -1,34 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2011-06-06 18:45+0800\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0\n"
-
-#: build/lib.linux-x86_64-2.7/online_status/templatetags/status_format.py:10
-#: online_status/templatetags/status_format.py:10
-msgid "online"
-msgstr "線上"
-
-#: build/lib.linux-x86_64-2.7/online_status/templatetags/status_format.py:12
-#: online_status/templatetags/status_format.py:12
-msgid "idle"
-msgstr "閒置"
-
-#: build/lib.linux-x86_64-2.7/online_status/templatetags/status_format.py:14
-#: online_status/templatetags/status_format.py:14
-msgid "offline"
-msgstr "離線"
