--- portal/middleware/django_online_status/online_status/__init__.py
@@ -1,7 +1,0 @@
-VERSION = (0, 1, 1)
-
-def get_version():
-    version = '%s.%s' % (VERSION[0], VERSION[1])
-    if VERSION[2]:
-        version = '%s.%s' % (version, VERSION[2])
-    return version
