--- portal/middleware/django_online_status/online_status/templates/online_status/user_status.html
@@ -1,2 +1,0 @@
-{% load status_format %}
-{{ onlinestatus.status|status_format:"text" }}
